Downloads/Kwanpa-Health-Service/temp_requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
alembic==1.17.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asyncpg==0.31.0
attrs==25.4.0
azure-ai-formrecognizer==3.3.3
azure-ai-vision-imageanalysis==1.0.0
azure-cognitiveservices-vision-computervision==0.9.1
azure-common==1.1.28
azure-core==1.36.0
azure-identity==1.25.1
azure-iot-hub==2.7.0
azure-storage-blob==12.27.1
bcrypt==4.0.1
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.120.0
frozenlist==1.8.0
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
isodate==0.7.2
Jinja2==3.1.6
jiter==0.12.0
Mako==1.3.10
MarkupSafe==3.0.3
msal==1.34.0
msal-extensions==1.3.1
msrest==0.7.1
multidict==6.7.0
oauthlib==3.3.1
openai==2.9.0
packaging==25.0
passlib==1.7.4
pillow==12.0.0
propcache==0.4.1
psycopg2==2.9.11
psycopg2-binary==2.9.11
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-http-client==3.3.7
python-jose==3.5.0
python-magic==0.4.27
python-multipart==0.0.20
PyYAML==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
sendgrid==6.12.5
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
tqdm==4.67.1
twilio==9.8.8
typing-inspection==0.4.2
typing_extensions==4.15.0
uamqp==1.6.11
urllib3==2.6.1
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
Werkzeug==3.1.4
yarl==1.22.0
